Time: 0 ms (100%), Space: 43 MB (79.94%) - LeetHub

+++ 0098-validate-binary-search-tree/0098-validate-binary-search-tree.java
@@ -17,10 +17,10 @@ class Solution {
     public boolean isValidBST(TreeNode root) {
         if(root == null)
             return false;
-        return isBST(root, Integer.MIN_VALUE, Integer.MAX_VALUE);    
+        return isBST(root, Long.MIN_VALUE, Long.MAX_VALUE);    
     }
 
-    boolean isBST(TreeNode root, int minVal, int maxVal){
+    boolean isBST(TreeNode root, long minVal, long maxVal){
        if(root == null) return true;
        if(root.val >= maxVal || root.val <= minVal) return false;
        return isBST(root.left, minVal , root.val) && isBST(root.right, root.val, maxVal);
